build: (first draft, with 2 examples)- generates alt text and captions that automatically include specific quantities and parameters from the model results file.
The parameters are extracted from the model results file in create_figure_doc.R. The values for those parameters are substituted for placeholders in the pre-written captions and alt text (sourced from captions_alttext.csv). The parameters are exported to the global R environment and referenced in the preamble in the skeleton .qmd file.